Fix documentation code actions and invocation args context aware completions issues
