+++ language-server/modules/langserver-core/src/main/java/org/ballerinalang/langserver/command/CommandUtil.java
@@ -156,13 +156,11 @@ public class CommandUtil {
 
         if (isService) {
             CodeAction action = new CodeAction(CommandConstants.CREATE_TEST_SERVICE_TITLE);
-            action.setKind(CodeActionKind.Source);
             action.setCommand(new Command(CommandConstants.CREATE_TEST_SERVICE_TITLE,
                                           CreateTestExecutor.COMMAND, args));
             actions.add(action);
         } else if (isFunction) {
             CodeAction action = new CodeAction(CommandConstants.CREATE_TEST_FUNC_TITLE);
-            action.setKind(CodeActionKind.Source);
             action.setCommand(new Command(CommandConstants.CREATE_TEST_FUNC_TITLE,
                                           CreateTestExecutor.COMMAND, args));
             actions.add(action);
@@ -659,7 +657,6 @@ public class CommandUtil {
         CodeAction action = new CodeAction(CommandConstants.ADD_DOCUMENTATION_TITLE);
         action.setCommand(new Command(CommandConstants.ADD_DOCUMENTATION_TITLE,
                                       AddDocumentationExecutor.COMMAND, args));
-        action.setKind(CodeActionKind.Source);
         return action;
     }
 
@@ -668,7 +665,6 @@ public class CommandUtil {
         List<Object> args = new ArrayList<>(Collections.singletonList(docUriArg));
         CodeAction action = new CodeAction(CommandConstants.ADD_ALL_DOC_TITLE);
         action.setCommand(new Command(CommandConstants.ADD_ALL_DOC_TITLE, AddAllDocumentationExecutor.COMMAND, args));
-        action.setKind(CodeActionKind.Source);
         return action;
     }
 
@@ -679,7 +675,6 @@ public class CommandUtil {
         CodeAction codeAction = new CodeAction(CommandConstants.CREATE_INITIALIZER_TITLE);
         codeAction.setCommand(new Command(CommandConstants.CREATE_INITIALIZER_TITLE,
                                           CreateObjectInitializerExecutor.COMMAND, args));
-        codeAction.setKind(CodeActionKind.Source);
         return codeAction;
     }
